requirements.txt
certifi==2018.8.24
chardet==3.0.4
cx-Oracle==6.4.1
cycler==0.10.0
idna==2.7
inflection==0.3.1
kiwisolver==1.0.1
matplotlib==3.0.0
more-itertools==4.3.0
mpl-finance==0.10.0
mysql-connector-python==8.0.12
mysqlclient==1.3.13
numpy==1.15.1
pandas==0.23.4
pkg-resources==0.0.0
protobuf==3.6.1
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
Quandl==3.4.2
requests==2.19.1
six==1.11.0
urllib3==1.23
